# openapi/openapi.json
{
  "info": {
    "description": "REST API for managing technical processes, products, documents, and resources in the PSS application",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "Proprietary"
    },
    "title": "Tech Process Viewer API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:7239",
      "description": "Backend PSS REST API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-APL-SessionKey",
        "description": "Session key obtained from /api/connect endpoint"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "status": {
            "type": "string",
            "description": "Error name"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Errors",
            "additionalProperties": {}
          }
        }
      },
...
      "Error1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error description"
          },
          "status_code": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": {}
          }
        },
        "required": [
          "error"
        ]
      },
...
      "OrganizationListItem": {
        "type": "object",
        "properties": {
          "organization_id": {
            "type": "integer",
            "description": "Organization system ID"
          },
          "id": {
            "type": "string",
            "description": "Organization ID"
          },
          "name": {
            "type": "string",
            "description": "Organization name"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "nullable": true
          }
        }
      },
      "OrganizationListResponse": {
        "type": "object",
        "properties": {
          "organizations": {
            "type": "array",
            "description": "List of organizations",
            "items": {
              "$ref": "#/components/schemas/OrganizationListItem"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of organizations"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "per_page": {
            "type": "integer",
            "description": "Items per page"
          }
        }
      },
      "OrganizationCreate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Organization ID"
          },
          "name": {
            "type": "string",
            "description": "Organization name"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "OrganizationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Organization data",
            "additionalProperties": {}
          },
          "organization_id": {
            "type": "integer",
            "description": "Organization system ID"
          }
        },
        "required": [
          "success"
        ]
      },
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Organization ID"
          },
          "name": {
            "type": "string",
            "description": "Organization name"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "nullable": true
          },
          "customized": {
            "type": "boolean",
            "description": "Whether organization is customized",
            "nullable": true
          }
        }
      },
      "OrganizationDetail": {
        "type": "object",
        "properties": {
          "sys_id": {
            "type": "integer",
            "description": "System ID"
          },
          "type": {
            "type": "string",
            "description": "Entity type"
          },
          "access": {
            "type": "string",
            "description": "Access level"
          },
          "attributes": {
            "description": "Organization attributes",
            "allOf": [
              {
                "$ref": "#/components/schemas/Organization"
              }
            ]
          }
        }
      },
      "OrganizationUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Description",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "UNPROCESSABLE_ENTITY": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "DEFAULT_ERROR": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "SessionKey": []
    }
  ],
  "paths": {
...
    "/api/v1/organizations/": {
      "get": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "List all organizations",
        "tags": [
          "organizations"
        ],
        "description": "Get list of organizations with optional filtering"
      },
      "post": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreate"
              }
            }
          }
        },
        "summary": "Create a new organization",
        "tags": [
          "organizations"
        ],
        "description": "Create a new organization"
      }
    },
    "/api/v1/organizations/{organization_id}": {
      "get": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetail"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Get organization by ID",
        "tags": [
          "organizations"
        ],
        "description": "Get organization details by ID"
      },
      "put": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdate"
              }
            }
          }
        },
        "summary": "Update organization",
        "tags": [
          "organizations"
        ],
        "description": "Update organization"
      },
      "delete": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Delete organization",
        "tags": [
          "organizations"
        ],
        "description": "Delete organization"
      },
      "parameters": [
        {
          "in": "path",
          "name": "organization_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
...
  },
...
  "openapi": "3.0.3"
}